+++ src/fastoad/module_management/tests/data/dummy_services/dummy_services.py
@@ -1,5 +1,5 @@
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -11,7 +11,10 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from fastoad.module_management.tests.test_service_registry import RegisterDummyServiceA, DummyBase
+from fastoad.module_management.tests.test_register_specialized_service import (
+    DummyBase,
+    RegisterDummyServiceA,
+)
 
 
 @RegisterDummyServiceA("dummy.provider.1")
