+++ src/fastoad/module_management/tests/test_register_specialized_service.py
